Move site_limit function to Billing.Quota
